+++ mqtt/mqtt-broker/src/lib.rs
@@ -12,12 +12,12 @@
 )]
 
 mod broker;
-mod configuration;
 mod connection;
 mod error;
 mod persist;
 mod server;
 mod session;
+mod settings;
 mod snapshot;
 mod state_change;
 mod subscription;
@@ -35,7 +35,6 @@ use mqtt3::proto;
 use mqtt_broker_core::{auth::AuthId, ClientId};
 
 pub use crate::broker::{Broker, BrokerBuilder, BrokerHandle};
-pub use crate::configuration::{BrokerConfig, SessionConfig};
 pub use crate::connection::ConnectionHandle;
 pub use crate::error::{DetailedErrorValue, Error, InitializeBrokerError};
 pub use crate::persist::{
@@ -43,6 +42,7 @@ pub use crate::persist::{
 };
 pub use crate::server::Server;
 pub use crate::session::SessionState;
+pub use crate::settings::{BrokerConfig, SessionConfig};
 pub use crate::snapshot::{
     BrokerSnapshot, SessionSnapshot, ShutdownHandle, Snapshotter, StateSnapshotHandle,
 };
