+++ mqtt/mqtt-broker/src/session/state/queue.rs
@@ -2,7 +2,7 @@ use std::{collections::VecDeque, num::NonZeroUsize};
 
 use mqtt3::proto;
 
-use crate::configuration::QueueFullAction;
+use crate::settings::QueueFullAction;
 
 /// `BoundedQueue` is a queue of publications with bounds by count and total payload size in bytes.
 ///
